backend/requirements.txt
fastapi
uvicorn
sqlmodel
SQLAlchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-jose
python-multipart
alembic
psycopg2-binary
